requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.0.9
cached-property==1.5.2
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
ChemDataExtractor==1.3.0
click==8.1.7
comm==0.2.2
cryptography==43.0.1
cssselect==1.2.0
dataclasses-json==0.6.7
DAWG==0.8.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
fqdn==1.5.1
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.2
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter_client==8.6.2
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.117
lxml==5.3.0
MarkupSafe==2.1.3
marshmallow==3.22.0
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.24.4
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.4
pdfminer.six==20240706
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
pkgutil_resolve_name==1.3.10
platformdirs==4.3.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==5.7.0
ptyprocess==0.7.0
PubChemPy==1.0.4
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PySocks==1.7.1
python-crfsuite==0.9.10
python-dateutil==2.9.0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==5.3
pyzmq==25.1.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
Send2Trash==1.8.3
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.34
stack-data==0.6.2
tenacity==8.5.0
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-utils==0.1.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
widgetsnbextension==4.0.13
yarl==1.11.1
zipp==3.20.1
zstandard==0.19.0
